+++ src/index.ts
@@ -1,3 +1,6 @@
 export * from './add';
+export * from './divide';
+export * from './mean';
+export * from './meanBy';
 export * from './sum';
 export * from './sumBy';