+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.2.0 - 2022-09-14
+
+- `divide()` function
+- `mean()` function
+- `meanBy()` function
+- remove lodash-es dependency
+
 ## 1.1.0 - 2022-09-14
 
 - `sumBy()` function